Test cleanup and adding tests of other projects
Other projects could be affected by can-jquery's behavior.
This runs the tests for those projects alongside the tests for can-jquery.
