+++ test/test.js
@@ -1,3 +1,9 @@
-import '../can-jquery_test';
-import '../can-jquery-stache-bindings_test';
-import '../legacy_test';
+require('can-jquery');
+
+require('../can-jquery_test');
+
+require('can-component/test/component-test');
+require('can-control/can-control_test');
+require('can-stache/test/stache-test');
+require('can-stache-bindings/test/bindings-test');
+require('can-util/test/test');
